Update Player.recordShot, Fleet.checkShot - checkShot returns a boolean array of length 2, returning if a boat is hit and if so, if the boat got destroyed
Before, gotDestroyed returned true as long as any boat was destroyed, not the boat which was hit with the current shot. Now Fleet correctly returns true if boat hit, true if boat destoryed

Fleet no longer uses aCoordinatesUsed, since this List is not updated upon shots, it should only be used for placing ships and checking overlapping ships

todo: Fleet does not return which BoatType or BoatTypeName. This could be improved by either Fleet returning a List which also contains the destroyed boats name